compat

+++ pandas/tests/series/test_ufunc.py
@@ -41,7 +41,7 @@ def test_unary_ufunc(ufunc, sparse):
     array = np.random.randint(0, 10, 10, dtype='int64')
     array[::2] = 0
     if sparse:
-        array = pd.SparseArray(array, dtype=pd.SparseDtype('int', 0))
+        array = pd.SparseArray(array, dtype=pd.SparseDtype('int64', 0))
 
     index = list(string.ascii_letters[:10])
     name = "name"
@@ -59,8 +59,8 @@ def test_binary_ufunc_with_array(flip, sparse, ufunc, arrays_for_binary_ufunc):
     # Test that ufunc(Series(a), array) == Series(ufunc(a, b))
     a1, a2 = arrays_for_binary_ufunc
     if sparse:
-        a1 = pd.SparseArray(a1, dtype=pd.SparseDtype('int', 0))
-        a2 = pd.SparseArray(a2, dtype=pd.SparseDtype('int', 0))
+        a1 = pd.SparseArray(a1, dtype=pd.SparseDtype('int64', 0))
+        a2 = pd.SparseArray(a2, dtype=pd.SparseDtype('int64', 0))
 
     name = "name"  # op(Series, array) preserves the name.
     series = pd.Series(a1, name=name)
@@ -87,8 +87,8 @@ def test_binary_ufunc_with_index(flip, sparse, ufunc, arrays_for_binary_ufunc):
     #   * ufunc(Index, Series) dispatches to Series (returns a Series)
     a1, a2 = arrays_for_binary_ufunc
     if sparse:
-        a1 = pd.SparseArray(a1, dtype=pd.SparseDtype('int', 0))
-        a2 = pd.SparseArray(a2, dtype=pd.SparseDtype('int', 0))
+        a1 = pd.SparseArray(a1, dtype=pd.SparseDtype('int64', 0))
+        a2 = pd.SparseArray(a2, dtype=pd.SparseDtype('int64', 0))
 
     name = "name"  # op(Series, array) preserves the name.
     series = pd.Series(a1, name=name)
@@ -118,8 +118,8 @@ def test_binary_ufunc_with_series(flip, shuffle, sparse, ufunc,
     #   with alignment between the indices
     a1, a2 = arrays_for_binary_ufunc
     if sparse:
-        a1 = pd.SparseArray(a1, dtype=pd.SparseDtype('int', 0))
-        a2 = pd.SparseArray(a2, dtype=pd.SparseDtype('int', 0))
+        a1 = pd.SparseArray(a1, dtype=pd.SparseDtype('int64', 0))
+        a2 = pd.SparseArray(a2, dtype=pd.SparseDtype('int64', 0))
 
     name = "name"  # op(Series, array) preserves the name.
     series = pd.Series(a1, name=name)
@@ -192,8 +192,8 @@ def test_multiple_ouput_binary_ufuncs(ufunc, sparse, shuffle,
     a2[a2 == 0] = 1
 
     if sparse:
-        a1 = pd.SparseArray(a1, dtype=pd.SparseDtype('int', 0))
-        a2 = pd.SparseArray(a2, dtype=pd.SparseDtype('int', 0))
+        a1 = pd.SparseArray(a1, dtype=pd.SparseDtype('int64', 0))
+        a2 = pd.SparseArray(a2, dtype=pd.SparseDtype('int64', 0))
 
     s1 = pd.Series(a1)
     s2 = pd.Series(a2)
